ChromaHacker/pyproject.toml
[tool.poetry]
name = "chromahacker"
version = "0.3.1"
description = ""
authors = ["FaunuX <judahpspiegel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.7.0.72"
pillow = "^9.5.0"
requests = "^2.31.0"

[tool.poetry.scripts]
test = "tests.main:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
